% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bmsPage}
\alias{bmsPage}
\title{Create a BMS-themed Page}
\usage{
bmsPage(..., title = NULL, theme = bms_theme())
}
\arguments{
\item{...}{Content of the page, \code{\link[=sidebarItem]{sidebarItem()}}}

\item{title}{Title of the page.}

\item{theme}{\code{\link[bslib:bs_theme]{bslib::bs_theme()}}}
}
\description{
Create a BMS-theme shiny page.
}
