% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonText.R
\name{buttonTextInput}
\alias{buttonTextInput}
\title{Button Text}
\usage{
buttonTextInput(
  inputId,
  btn_content = "Create",
  placeholder = "Title",
  validate_content = shiny::icon("paper-plane"),
  value = ""
)
}
\arguments{
\item{inputId}{Id of input.}

\item{btn_content}{Content of initial button.}

\item{placeholder}{Placeholder for revealed text input.}

\item{validate_content}{Content of revealed button.}

\item{value}{Initial value.}
}
\description{
A button that "transforms" into a text input.
}
\examples{
library(shiny)

ui <- fluidPage(
 buttonTextInput("theId", 0)
)

server <- function(input, output){

 observeEvent(input$theId, {
   print(input$theId)
 })

}

if(interactive())
 shinyApp(ui, server)

}
