% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-classes.R
\name{constructSidebar}
\alias{constructSidebar}
\title{Sidebar Constructor}
\usage{
constructSidebar(x)
}
\arguments{
\item{x}{Object to add class to.}
}
\description{
Sidebar constructor
}
\keyword{internal}
