% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-classes.R
\name{constructSidebarCollapsible}
\alias{constructSidebarCollapsible}
\title{Sidebar Collapsible Constructor}
\usage{
constructSidebarCollapsible(x)
}
\arguments{
\item{x}{Object to add class to.}
}
\description{
Sidebar collapsible constructor
}
\keyword{internal}
