% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-classes.R
\name{constructSidebarCollapsibleItem}
\alias{constructSidebarCollapsibleItem}
\title{Sidebar Collapsible Item Constructor}
\usage{
constructSidebarCollapsibleItem(x)
}
\arguments{
\item{x}{Object to add class to.}
}
\description{
Sidebar collapsible item constructor
}
\keyword{internal}
