% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-classes.R
\name{constructSidebarItem}
\alias{constructSidebarItem}
\title{Sidebar Item Constructor}
\usage{
constructSidebarItem(x)
}
\arguments{
\item{x}{Object to add class to.}
}
\description{
Sidebar item constructor
}
\keyword{internal}
