% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-server.R
\name{insert_sidebar_collapsible}
\alias{insert_sidebar_collapsible}
\title{Insert Sidebar Collapsible}
\usage{
insert_sidebar_collapsible(title, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title}{Title of the item.}

\item{session}{A valid Shiny session.}
}
\description{
Insert a sidebar collapsible, similar to tab but works
with a sidebar.
}
