% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-server.R
\name{insert_sidebar_collapsible_item}
\alias{insert_sidebar_collapsible_item}
\title{Insert Sidebar Collapsible Item}
\usage{
insert_sidebar_collapsible_item(
  target,
  title,
  ...,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{target}{Title of \code{\link[=sidebarCollapsible]{sidebarCollapsible()}} under which to place this entry.}

\item{title}{Title of the item.}

\item{...}{Content of the item.}

\item{session}{A valid Shiny session.}
}
\description{
Insert a sidebar collapsible, similar to tab but works
with a sidebar.
}
