% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{insertSidebarItem}
\alias{insertSidebarItem}
\title{Sidebar Item}
\usage{
insertSidebarItem(title, ..., session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title}{Title of the item.}

\item{...}{Content of the item.}

\item{session}{A valid Shiny session.}
}
\description{
Create a sidebar item, similar to tab but works
with a sidebar.
Meant to be passed to \code{\link[=sidebar]{sidebar()}}.
}
