% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbarItem}
\alias{navbarItem}
\title{Navbar Item}
\usage{
navbarItem(title, ...)
}
\arguments{
\item{title}{Title of navbar item.}

\item{...}{Any of attributes.}
}
\description{
Create a navbar item.
Meant to be passed to \code{\link[=navbar]{navbar()}}.
}
