% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{not_missing_failure}
\alias{not_missing_failure}
\title{Is not Missing Assertion}
\arguments{
\item{call}{Used to deparse expression.}

\item{env}{Environment.}
}
\description{
Is not Missing Assertion
}
\keyword{internal}
