% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-server.R
\name{select_sidebar_item}
\alias{select_sidebar_item}
\title{Insert Sidebar Item}
\usage{
select_sidebar_item(title, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{title}{Title of the item.}

\item{session}{A valid Shiny session.}
}
\description{
Insert a sidebar item, similar to tab but works
with a sidebar.
}
