% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebar}
\alias{sidebar}
\title{Sidebar}
\usage{
sidebar(..., title = NULL)
}
\arguments{
\item{...}{Content of the sidebar.}

\item{title}{Title of the sidebar.}
}
\description{
Create a sidebar.
Meant to be passed to \code{\link[=bmsPage]{bmsPage()}}.
}
