% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarCollapsible}
\alias{sidebarCollapsible}
\title{Sidebar Collapsible}
\usage{
sidebarCollapsible(title, ...)
}
\arguments{
\item{title}{Title of the item.}

\item{...}{Content of the item.}
}
\description{
Create a sidebar collapsible, similar to tab but works
with a sidebar.
Meant to be passed to \code{\link[=sidebar]{sidebar()}}.
}
