% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarCollapsibleItem}
\alias{sidebarCollapsibleItem}
\title{Sidebar Collapsible Item}
\usage{
sidebarCollapsibleItem(title, ...)
}
\arguments{
\item{title}{Title of the item.}

\item{...}{Content of the item.}
}
\description{
Create a sidebar collapsible item, similar to tab but works
with a sidebar.
Meant to be passed to \code{\link[=sidebarCollapsible]{sidebarCollapsible()}}.
}
