% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarItemAny}
\alias{sidebarItemAny}
\title{Sidebar Item Any}
\usage{
sidebarItemAny(...)
}
\arguments{
\item{...}{Content of the item.}
}
\description{
Create a sidebar item with any content,
does not trigger the opening of a tab.
}
