% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar-classes.R
\name{sidebarItemGeneric}
\alias{sidebarItemGeneric}
\title{Sidebar Item Generic}
\usage{
sidebarItemGeneric(title, ..., description = NULL)
}
\arguments{
\item{x}{Object to add class to.}
}
\description{
Sidebar item generic.
}
\note{
It's not \code{generic} in the computing sense of the term.
}
\keyword{internal}
