% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttonText.R
\name{togglerTextInput}
\alias{togglerTextInput}
\title{Toggler Text Input}
\usage{
togglerTextInput(
  inputId,
  content = "Create",
  placeholder = "Title",
  validate_content = shiny::icon("paper-plane"),
  value = "",
  restore = FALSE
)
}
\arguments{
\item{inputId}{Id of input.}

\item{content}{Content of initial element.}

\item{placeholder}{Placeholder for revealed text input.}

\item{validate_content}{Content of revealed button.}

\item{value}{Initial value.}

\item{restore}{If \code{TRUE} it restores the text content to the
value set in the text input, if \code{FALSE} it restores the
to the \code{content}.}
}
\description{
An element that "transforms" into a text input.
}
\examples{
library(shiny)

ui <- fluidPage(
 togglerTextInput("theId", "Button")
)

server <- function(input, output){

 observeEvent(input$theId, {
   print(input$theId)
 })

}

if(interactive())
 shinyApp(ui, server)

}
