% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{bmsPage}
\alias{bmsPage}
\title{Create a BMS-themed Page}
\usage{
bmsPage(..., sidebar, navbar = NULL, title = NULL, theme = bms_theme())
}
\arguments{
\item{...}{Content of the page, \code{\link[=sidebarItem]{sidebarItem()}}}

\item{sidebar}{Sidebar as returned by \code{\link[=sidebar]{sidebar()}}.}

\item{navbar}{Navbar as returned by \code{\link[=navbar]{navbar()}}.}

\item{title}{Title of the page.}

\item{theme}{\code{\link[bslib:bs_theme]{bslib::bs_theme()}}}
}
\description{
Create a BMS-theme shiny page.
}
