% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar}
\alias{navbar}
\title{Navbar}
\usage{
navbar(..., title = NULL, alignment = c("left", "right"))
}
\arguments{
\item{...}{Any number of \code{\link[=navbarItem]{navbarItem()}}.}

\item{title}{Title of navbar.}

\item{alignment}{Navbar alignment}
}
\description{
Create a navbar.
Meant to be passed to \code{\link[=bmsPage]{bmsPage()}}.
}
