% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Sidebar Item}
\usage{
sidebarItem(title, ..., description = NULL)
}
\arguments{
\item{title}{Title of the item.}

\item{...}{Content of the item.}

\item{description}{Description of item, shows insidebar}
}
\description{
Create a sidebar item, similar to tab but works
with a sidebar.
Meant to be passed to \code{\link[=sidebar]{sidebar()}}.
}
