% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebar.R
\name{sidebarItem}
\alias{sidebarItem}
\title{Sidebar Item}
\usage{
sidebarItem(title, ...)
}
\arguments{
\item{title}{Title of the sidebar.}

\item{...}{Content of the sidebar.}
}
\description{
Create a sidebar item, similar to tab but works
with a sidebar.
Meant to be passed to \code{\link[=bmsPage]{bmsPage()}}.
}
